feat(tests): move env-specific tests to dedicated e2e suite
- extract zsh binary/env tests into tests/e2e.rs, all marked #[ignore]
- keep pure parser unit test in zsh_alias_scan.rs
- run e2e suite explicitly in CI: cargo test --test e2e -- --ignored
- add apt-get update before zsh install on Linux runners
